Extract sharing and embedding actions
