+++ frontend/src/metabase/dashboard/actions/actions.js
@@ -124,18 +124,6 @@ export const setShowLoadingCompleteFavicon = createAction(
   SET_SHOW_LOADING_COMPLETE_FAVICON,
 );
 
-export const setSharing = isSharing => dispatch => {
-  if (isSharing) {
-    dispatch(
-      setSidebar({
-        name: SIDEBAR_NAME.sharing,
-      }),
-    );
-  } else {
-    dispatch(closeSidebar());
-  }
-};
-
 export const showClickBehaviorSidebar = dashcardId => dispatch => {
   if (dashcardId != null) {
     dispatch(
@@ -775,38 +763,6 @@ export const fetchDashboard = createThunkAction(
         parameterValues: parameterValuesById,
       };
     };
-  },
-);
-
-export const UPDATE_ENABLE_EMBEDDING =
-  "metabase/dashboard/UPDATE_ENABLE_EMBEDDING";
-export const updateEnableEmbedding = createAction(
-  UPDATE_ENABLE_EMBEDDING,
-  ({ id }, enable_embedding) => DashboardApi.update({ id, enable_embedding }),
-);
-
-export const UPDATE_EMBEDDING_PARAMS =
-  "metabase/dashboard/UPDATE_EMBEDDING_PARAMS";
-export const updateEmbeddingParams = createAction(
-  UPDATE_EMBEDDING_PARAMS,
-  ({ id }, embedding_params) => DashboardApi.update({ id, embedding_params }),
-);
-
-export const CREATE_PUBLIC_LINK = "metabase/dashboard/CREATE_PUBLIC_LINK";
-export const createPublicLink = createAction(
-  CREATE_PUBLIC_LINK,
-  async ({ id }) => {
-    const { uuid } = await DashboardApi.createPublicLink({ id });
-    return { id, uuid };
-  },
-);
-
-export const DELETE_PUBLIC_LINK = "metabase/dashboard/DELETE_PUBLIC_LINK";
-export const deletePublicLink = createAction(
-  DELETE_PUBLIC_LINK,
-  async ({ id }) => {
-    await DashboardApi.deletePublicLink({ id });
-    return { id };
   },
 );
