+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -1,5 +1,6 @@
 export * from "./actions";
 export * from "./core";
 export * from "./parameters";
+export * from "./sharing";
 export * from "./ui";
 export * from "./writeback";
